requirements.txt
pandas
regex
requests
tqdm
torch      # CPU-сборка автоматически подтянется
open_clip_torch
Pillow
